Create ifstatement
Complete the code in the editor to find the average rating for free apps.

Inside the for loop, complete the following:

Assign the price of an app as a float to a variable named price. The price is the fifth element in each row (the index starts at 0).
If price == 0.0, append the value stored in rating to the free_apps_ratings list using the list_name.append() function (note the free_apps_ratings is already defined in the code editor). Be careful with indentation.
Outside the for loop body, compute the average rating of free apps. Assign the result to a variable named avg_rating_free. The ratings are stored in the free_apps_ratings list.